.mcp/requirements.txt
# MCP Server 依赖
mcp>=1.0.0
httpx>=0.27.0
Pillow>=10.0.0
